Replace Skia with OpenGL
Skia was a heavy dependency -- undocumented, unintuitive to set up.
It used OpenGL under the hood. The graphics requirement was not so
sophisticated, so most of Skia's features were left unused. Since
the codebase is large, it consumed a signficant amount of compile
time and also increased binary size.

It appears Skia used 6/8x MSAA; we use 2x.