Use PipeWire as the audio backend.
With SDL 2/3 it was not possible to monitor system audio, which is desired. Though SDL provides a hint for capturing monitor devices, it is exclusive to PulseAudio. PipeWire supports capturing monitor devices, and SDL also uses PipeWire as its backend. So it's only better, for our purpose, to directly use PipeWire.